Add files via upload
Hello readers, To create a Fully Responsive Personal Portfolio Website using HTML CSS & JavaScript, web design stuff like [Navigation Bar, Cards, Image Slider, Owl Carousel, Buttons, and many more] but still, I haven’t created a full website and now it’s time to create a portfolio website.
Personal portfolio sites are consistent that need to be taken care of throughout your work. It gives a convenient way for potential clients to view your work while also letting you expand on your skills/experiences and services.
On this site [Personal Portfolio], there are five sections on one page – Home, About, Services, Skills, and Contact, and each section is attractive and eye-catching. On the home page of this site, on the top, there is a sticky navigation bar with a logo on the left side and some navigation links are on the right side.
On the left side of the home page, there are texts which are about the author’s name, profession, and a button labeled “Hire me” as you can see in the image.

To create this website (Personal Portfolio Website), you need to create three Files: HTML File, CSS File, and JavaScript File. First, create an HTML file with the name of index.html. Remember, you’ve to create a file with .html extension.